fix(tests): resolve cross-file mock leakage breaking CI
bun:test mock.module mutates the global ES registry and persists across
files. Partial mocks that omitted exports used by other tests caused
TypeErrors like "execute is not a function" and "guildExists is not a
function" when test files ran together in CI (serialized differently
than local).

- Extract setActivityCore into its own module so worker/index.test.ts
  can mock the setActivity.js wrapper without clobbering setActivityCore
  imports used by setActivity.test.ts.
- Fill out partial mocks to include every export the real module
  provides (guildDatabase: guildExists + pruneGuilds + ensureGuildExists;
  command modules: default + named execute + slashCommand).
