+++ tests/commands/setup/channel.test.ts
@@ -13,7 +13,11 @@ describe("setup channel command", () => {
 
     mock.module("../../../src/utils/logger.js", () => ({ default: logger }));
     mock.module("../../../src/database/index.js", () => ({ db, queryClient: () => Promise.resolve([]) }));
-    mock.module("../../../src/utils/guildDatabase.js", () => ({ guildExists: sinon.stub().resolves(true) }));
+    mock.module("../../../src/utils/guildDatabase.js", () => ({
+      guildExists: sinon.stub().resolves(true),
+      pruneGuilds: sinon.stub().resolves(),
+      ensureGuildExists: sinon.stub().resolves(),
+    }));
 
     const mod = await import("../../../src/commands/setup/channel.js");
