+++ tests/events/interactionCreate.test.ts
@@ -9,19 +9,21 @@ const setupAutocomplete = sinon.stub().resolves();
 
 // Top-level mocks to avoid cross-file interference
 mock.module("../../src/utils/logger.js", () => ({ default: logger }));
-mock.module("../../src/commands/help.js", () => ({ default: { execute: executeStub } }));
-mock.module("../../src/commands/about.js", () => ({ default: { execute: executeStub } }));
-mock.module("../../src/commands/changelog.js", () => ({ default: { execute: executeStub } }));
-mock.module("../../src/commands/quote.js", () => ({ default: { execute: executeStub } }));
-mock.module("../../src/commands/suggestion.js", () => ({ default: { execute: executeStub } }));
-mock.module("../../src/commands/invite.js", () => ({ default: { execute: executeStub } }));
-mock.module("../../src/commands/admin/index.js", () => ({ default: { execute: executeStub } }));
+mock.module("../../src/commands/help.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "help" } }));
+mock.module("../../src/commands/about.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "about" } }));
+mock.module("../../src/commands/changelog.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "changelog" } }));
+mock.module("../../src/commands/quote.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "quote" } }));
+mock.module("../../src/commands/suggestion.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "suggestion" } }));
+mock.module("../../src/commands/invite.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "invite" } }));
+mock.module("../../src/commands/admin/index.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "admin" } }));
 mock.module("../../src/commands/setup/index.js", () => ({
   default: { execute: executeStub },
+  execute: executeStub,
   setupAutocomplete,
+  slashCommand: { name: "setup" },
 }));
-mock.module("../../src/commands/premium.js", () => ({ default: { execute: executeStub } }));
-mock.module("../../src/commands/owner/index.js", () => ({ default: { execute: executeStub } }));
+mock.module("../../src/commands/premium.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "premium" } }));
+mock.module("../../src/commands/owner/index.js", () => ({ default: { execute: executeStub }, execute: executeStub, slashCommand: { name: "owner" } }));
 
 const { interactionCreateEvent } = await import("../../src/events/interactionCreate.js");
